requirements.txt
flask
flask-cors
pandas
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
chromadb
